git-svn-id: http://eathena.googlecode.com/svn/branches/stable@11665 9c4de8d1-4af9-ca9b-867f-b47ac08f6c80

+++ src/map/unit.c
@@ -661,7 +661,7 @@ int unit_stop_walking(struct block_list *bl,int type)
 	return 1;
 }
 
-int unit_skilluse_id(struct block_list *src, int target_id, int skill_num, int skill_lv)
+int unit_skilluse_id(struct block_list *src, int target_id, short skill_num, short skill_lv)
 {
 	if(skill_num < 0) return 0;
 
@@ -775,7 +775,7 @@ int unit_set_walkdelay(struct block_list *bl, unsigned int tick, int delay, int 
 	return 1;
 }
 
-int unit_skilluse_id2(struct block_list *src, int target_id, int skill_num, int skill_lv, int casttime, int castcancel)
+int unit_skilluse_id2(struct block_list *src, int target_id, short skill_num, short skill_lv, int casttime, int castcancel)
 {
 	struct unit_data *ud;
 	struct status_data *tstatus;
@@ -1048,7 +1048,7 @@ int unit_skilluse_id2(struct block_list *src, int target_id, int skill_num, int 
 	return 1;
 }
 
-int unit_skilluse_pos(struct block_list *src, short skill_x, short skill_y, int skill_num, int skill_lv)
+int unit_skilluse_pos(struct block_list *src, short skill_x, short skill_y, short skill_num, short skill_lv)
 {
 	if(skill_num < 0)
 		return 0;
@@ -1059,7 +1059,7 @@ int unit_skilluse_pos(struct block_list *src, short skill_x, short skill_y, int 
 	);
 }
 
-int unit_skilluse_pos2( struct block_list *src, short skill_x, short skill_y, int skill_num, int skill_lv, int casttime, int castcancel)
+int unit_skilluse_pos2( struct block_list *src, short skill_x, short skill_y, short skill_num, short skill_lv, int casttime, int castcancel)
 {
 	struct map_session_data *sd = NULL;
 	struct unit_data        *ud = NULL;
